Add settings for IOTA/SOTA/POTA mode and location, and macro set We have two new fields in sbitx/sbitxd: xota_loc/LOCATION (text) and xota/xOTA (selections NONE/IOTA/SOTA/POTA). These can now be set via the front panel settings page, which has modified layout to make everything fit nicely, with smaller margins.
The actual behavior in these modes can be refined, but probably
this is a good enough way to handle the settings.

The #current_macro/MACRO field was already there. On my radio I added
/usr/local/share/sbitxd/web/SOTA.mc which can call CQ SOTA instead of
simply CQ. I haven't bothered to tie together the macro and xOTA
settings. All kinds of "special CQ" words are possible. It's up to the
sbitx/sbitxd software to find the macro files, but here we unfortunately
also have to hard-code the choices. The protocol should actually be
modified so that the radio sends the choices to the front panel, but so
far it's not that way. You can choose them with the rotary encoder, as
with the other FIELD_SELECTION fields.